moved 3 pages into seprate files
