+++ package.json
@@ -10,12 +10,15 @@
     "@types/node": "^16.18.76",
     "@types/react": "^18.2.48",
     "@types/react-dom": "^18.2.18",
+    "axios": "^1.6.8",
     "bootstrap": "^5.3.2",
+    "dotenv": "^16.4.5",
     "gh-pages": "^6.1.1",
-    "openai": "^4.26.0",
+    "openai": "^4.38.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.1",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
